# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationResultList": {
        "description": "Recent persisted judged answer results.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResultSummary"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "title": "EvaluationResultList",
        "type": "object"
      },
...
      "EvaluationRunList": {
        "description": "Recent persisted answer-evaluation runs.",
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "title": "EvaluationRunList",
        "type": "object"
      },
      "EvaluationRunStatus": {
        "description": "Lifecycle state for a persisted evaluation batch.",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "EvaluationRunStatus",
        "type": "string"
      },
...
      "EvaluationSourceType": {
        "description": "Sources that can feed an answer-quality evaluation run.",
        "enum": [
          "dataset",
          "monitored_runs"
        ],
        "title": "EvaluationSourceType",
        "type": "string"
      },
...
      "FeedbackEvent": {
        "description": "A persisted feedback event.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "feedback_id": {
            "minLength": 1,
            "title": "Feedback Id",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "run_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "useful": {
            "title": "Useful",
            "type": "boolean"
          }
        },
        "required": [
          "session_id",
          "useful",
          "submitted_at"
        ],
        "title": "FeedbackEvent",
        "type": "object"
      },
      "FeedbackRequest": {
        "description": "A user feedback submission for one browser session or returned run.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "request_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "run_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "useful": {
            "title": "Useful",
            "type": "boolean"
          }
        },
        "required": [
          "useful"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonitoringFeedbackFilter": {
        "description": "Feedback filters supported by the monitoring run list.",
        "enum": [
          "all",
          "useful",
          "not_useful",
          "none"
        ],
        "title": "MonitoringFeedbackFilter",
        "type": "string"
      },
...
      "MonitoringRunList": {
        "description": "Recent persisted monitoring runs.",
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/MonitoringRunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "title": "MonitoringRunList",
        "type": "object"
      },
...
      "RunKind": {
        "description": "The persisted run categories shown in monitoring dashboards.",
        "enum": [
          "direct",
          "agentic"
        ],
        "title": "RunKind",
        "type": "string"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Repo Deep Research"
    },
    "description": "Repo Deep Research indexes Python repositories, answers direct and agentic research questions with cited repository evidence, records monitoring data, and exposes persisted answer-evaluation results.",
    "license": {
      "name": "MIT"
    },
    "summary": "Evidence-grounded research for Python repositories.",
    "title": "Repo Deep Research",
    "version": "0.5.7"
  },
  "openapi": "3.1.0",
  "paths": {
    "/evaluations/results": {
      "get": {
        "operationId": "list_evaluation_results",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationSourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "run_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Kind"
            }
          },
          {
            "in": "query",
            "name": "context_label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Context Label"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResultList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List evaluation results",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/evaluations/runs": {
      "get": {
        "operationId": "list_evaluation_runs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationSourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationRunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List evaluation runs",
        "tags": [
          "evaluations"
        ]
      }
    },
...
    "/feedback": {
      "post": {
        "operationId": "record_feedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Persist answer feedback",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/monitoring/runs": {
      "get": {
        "operationId": "list_monitoring_runs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Kind"
            }
          },
          {
            "in": "query",
            "name": "repository_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Name"
            }
          },
          {
            "in": "query",
            "name": "has_error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Error"
            }
          },
          {
            "in": "query",
            "name": "feedback",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MonitoringFeedbackFilter",
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List monitoring runs",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/monitoring/runs/{request_id}": {
      "get": {
        "operationId": "get_monitoring_run_detail",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get monitoring run detail",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/rag": {
      "post": {
        "operationId": "run_direct_rag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer with direct RAG",
        "tags": [
          "answers"
        ]
      }
    },
    "/repositories/ingest": {
      "post": {
        "operationId": "ingest_repository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest and index a repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/research": {
      "post": {
        "operationId": "run_agentic_research",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer with bounded agentic research",
        "tags": [
          "answers"
        ]
      }
    }
  },
...
}